Split FlashcardsStore cloud flows into focused extensions
