--- apps/ios/Flashcards/Flashcards/FlashcardsStore+Cloud.swift
@@ -1,1570 +1,0 @@
-import Foundation
-
-enum CloudBootstrapEligibilityError: LocalizedError {
-    case remoteWorkspaceIsNotEmpty
-
-    var errorDescription: String? {
-        switch self {
-        case .remoteWorkspaceIsNotEmpty:
-            return "Choose a new or empty workspace on this server before uploading the current local data."
-        }
-    }
-}
-
-private enum PersistedCloudStateReconciliationOutcome {
-    case continueSync(hasStoredCredentials: Bool, hasStoredGuestSession: Bool)
-    case stopSync
-}
-
-private let blockedCloudIdentityConflictCodes: Set<String> = [
-    "SYNC_INSTALLATION_PLATFORM_MISMATCH",
-    "SYNC_REPLICA_CONFLICT"
-]
-
-enum FlashcardsUITestResetState: String {
-    case localGuest = "local_guest"
-    case localGuestSeededManualReviewCard = "local_guest_seeded_manual_review_card"
-    case localGuestSeededAIReviewCard = "local_guest_seeded_ai_review_card"
-}
-
-private enum FlashcardsUITestSeedData {
-    static let manualReviewFrontText: String = "Smoke seeded manual review question"
-    static let manualReviewBackText: String = "Smoke seeded manual review answer"
-    static let aiReviewFrontText: String = "Smoke seeded AI review question"
-    static let aiReviewBackText: String = "Smoke seeded AI review answer"
-    static let aiReviewTag: String = "smoke-seeded-ai-review"
-}
-
-@MainActor
-extension FlashcardsStore {
-    /**
-     Clears all account-scoped local state so the next cloud link starts from a
-     fresh local database and a freshly generated sync device id.
-     
-     This reset is the boundary that intentionally breaks identity continuity
-     across logout and account deletion. The next guest cloud restore creates a
-     brand new server-side guest session with a new guest user/workspace, so we
-     never try to merge a future guest account into another linked account as if
-     it were the same pre-reset guest identity.
-     */
-    private func resetLocalStateForCloudIdentityChange() throws {
-        let database = try requireLocalDatabase(database: self.database)
-
-        self.reviewRuntime.cancelForAccountDeletion()
-        self.cloudRuntime.cancelForAccountDeletion()
-        try self.cloudRuntime.clearCredentials()
-        try self.dependencies.guestCredentialStore.clearGuestSession()
-        clearStoredReviewFilters(userDefaults: self.userDefaults)
-        self.userDefaults.removeObject(forKey: reviewNotificationPromptStateUserDefaultsKey)
-        self.userDefaults.removeObject(forKey: reviewNotificationSuccessfulReviewCountUserDefaultsKey)
-        self.userDefaults.removeObject(forKey: reviewNotificationLastActiveAtUserDefaultsKey)
-        self.userDefaults.removeObject(forKey: accountDeletionPendingUserDefaultsKey)
-        self.userDefaults.removeObject(forKey: aiChatExternalProviderConsentUserDefaultsKey)
-        self.cachedAIChatStore?.clearLocalHistory()
-        clearStoredAIChatHistories(userDefaults: self.userDefaults)
-        self.reviewRuntime = ReviewQueueRuntime(
-            initialSelectedReviewFilter: .allCards,
-            reviewSeedQueueSize: reviewSeedQueueSize,
-            reviewQueueReplenishmentThreshold: reviewQueueReplenishmentThreshold
-        )
-        self.applyReviewPublishedState(
-            reviewState: ReviewQueueRuntime.makeInitialPublishedState(selectedReviewFilter: .allCards)
-        )
-        self.notificationPermissionPromptState = makeDefaultNotificationPermissionPromptState()
-        self.isReviewNotificationPrePromptPresented = false
-        self.accountDeletionState = .hidden
-        self.accountDeletionSuccessMessage = nil
-        self.cloudSyncFastPollingUntil = nil
-        self.pendingReviewCardIds = []
-        self.reviewSubmissionFailure = nil
-        self.isAccountDeletionRunning = false
-        self.clearTransientBanners()
-        self.lastSuccessfulCloudSyncAt = nil
-        self.syncStatus = .idle
-        self.globalErrorMessage = ""
-        try database.resetForAccountDeletion()
-        try self.reload()
-    }
-
-    private func resetLocalStateIfLinkedUserDiffers(nextUserId: String) throws {
-        guard self.cloudSettings?.cloudState == .linked else {
-            return
-        }
-
-        guard let linkedUserId = self.cloudSettings?.linkedUserId, linkedUserId.isEmpty == false else {
-            return
-        }
-
-        if linkedUserId != nextUserId {
-            try self.resetLocalStateForCloudIdentityChange()
-        }
-    }
-
-    private func loadGuestSessionForCurrentConfiguration() throws -> StoredGuestCloudSession? {
-        guard let storedGuestSession = try self.dependencies.guestCredentialStore.loadGuestSession() else {
-            return nil
-        }
-
-        let configuration = try self.currentCloudServiceConfiguration()
-        if storedGuestSession.apiBaseUrl != configuration.apiBaseUrl
-            || storedGuestSession.configurationMode != configuration.mode {
-            try self.dependencies.guestCredentialStore.clearGuestSession()
-            return nil
-        }
-
-        return storedGuestSession
-    }
-
-    private func loadAuthenticatedCloudAccountSnapshot(
-        credentials: StoredCloudCredentials,
-        configuration: CloudServiceConfiguration
-    ) async throws -> CloudAccountSnapshot {
-        try await self.cloudRuntime.fetchCloudAccount(
-            verifiedContext: CloudVerifiedAuthContext(
-                apiBaseUrl: configuration.apiBaseUrl,
-                credentials: credentials
-            )
-        )
-    }
-
-    private func isAuthenticatedSilentRestoreEligible(
-        configuration: CloudServiceConfiguration,
-        hasStoredCredentials: Bool,
-        hasStoredGuestSession: Bool
-    ) throws -> Bool {
-        guard configuration.mode == .official else {
-            return false
-        }
-        guard hasStoredCredentials else {
-            return false
-        }
-        guard hasStoredGuestSession == false else {
-            return false
-        }
-        guard self.cloudSettings?.cloudState == .disconnected || self.cloudSettings?.cloudState == .linkingReady else {
-            return false
-        }
-
-        let database = try requireLocalDatabase(database: self.database)
-        return try database.isSafeForAuthenticatedSilentRestore()
-    }
-
-    private func selectedWorkspaceForAuthenticatedSilentRestore(
-        account: CloudAccountSnapshot
-    ) throws -> CloudWorkspaceSummary {
-        guard let selectedWorkspace = account.workspaces.first(where: { workspace in
-            workspace.isSelected
-        }) else {
-            throw LocalStoreError.validation("Authenticated cloud account is missing a selected workspace")
-        }
-
-        return selectedWorkspace
-    }
-
-    private func performAuthenticatedSilentRestore(
-        credentials: StoredCloudCredentials,
-        configuration: CloudServiceConfiguration,
-        trigger: CloudSyncTrigger
-    ) async throws {
-        let account = try await self.loadAuthenticatedCloudAccountSnapshot(
-            credentials: credentials,
-            configuration: configuration
-        )
-        let selectedWorkspace = try self.selectedWorkspaceForAuthenticatedSilentRestore(account: account)
-
-        try await self.finishCloudLink(
-            linkedSession: CloudLinkedSession(
-                userId: account.userId,
-                workspaceId: selectedWorkspace.workspaceId,
-                email: account.email,
-                configurationMode: configuration.mode,
-                apiBaseUrl: configuration.apiBaseUrl,
-                authorization: .bearer(credentials.idToken)
-            ),
-            trigger: trigger
-        )
-    }
-
-    private func restoreAuthenticatedCloudSessionAfterReinstall(
-        configuration: CloudServiceConfiguration,
-        trigger: CloudSyncTrigger
-    ) async throws {
-        do {
-            let credentials = try await self.refreshCloudCredentials(forceRefresh: false)
-            try await self.performAuthenticatedSilentRestore(
-                credentials: credentials,
-                configuration: configuration,
-                trigger: trigger
-            )
-            return
-        } catch {
-            if self.isCloudAuthorizationError(error) == false {
-                throw error
-            }
-        }
-
-        let refreshedCredentials = try await self.refreshCloudCredentials(forceRefresh: true)
-        try await self.performAuthenticatedSilentRestore(
-            credentials: refreshedCredentials,
-            configuration: configuration,
-            trigger: trigger
-        )
-    }
-
-    private func shouldResetLocalStateAfterAuthenticatedSilentRestoreFailure(error: Error) -> Bool {
-        if error is URLError {
-            return true
-        }
-        if error is CloudAuthError {
-            return true
-        }
-        if error is CloudSyncError {
-            return true
-        }
-
-        return false
-    }
-
-    /**
-     Prevents stored credentials from silently restoring a different cloud
-     account into local state that still belongs to the previous user.
-     */
-    @discardableResult
-    private func resetLocalStateIfStoredCredentialsBelongToDifferentUser(
-        credentials: StoredCloudCredentials,
-        configuration: CloudServiceConfiguration
-    ) async throws -> Bool {
-        guard self.cloudSettings?.cloudState == .linked else {
-            return false
-        }
-
-        let authenticatedAccount = try await self.loadAuthenticatedCloudAccountSnapshot(
-            credentials: credentials,
-            configuration: configuration
-        )
-        guard let linkedUserId = self.cloudSettings?.linkedUserId, linkedUserId.isEmpty == false else {
-            return false
-        }
-
-        if linkedUserId == authenticatedAccount.userId {
-            return false
-        }
-
-        try self.resetLocalStateForCloudIdentityChange()
-        return true
-    }
-
-    func currentCloudServiceConfiguration() throws -> CloudServiceConfiguration {
-        try loadCloudServiceConfiguration(
-            bundle: .main,
-            userDefaults: self.userDefaults,
-            decoder: self.decoder
-        )
-    }
-
-    func validateCustomCloudServer(customOrigin: String) async throws -> CloudServiceConfiguration {
-        let configuration = try makeCustomCloudServiceConfiguration(customOrigin: customOrigin)
-        try await self.cloudServiceConfigurationValidator.validate(configuration: configuration)
-        return configuration
-    }
-
-    func applyCustomCloudServer(configuration: CloudServiceConfiguration) throws {
-        if configuration.mode != .custom {
-            throw LocalStoreError.validation("Custom server configuration is required")
-        }
-
-        try self.switchCloudServer(override: CloudServerOverride(customOrigin: try requireCustomOrigin(configuration: configuration)))
-    }
-
-    func resetToOfficialCloudServer() throws {
-        try self.switchCloudServer(override: nil)
-    }
-
-    func sendCloudSignInCode(email: String) async throws -> CloudSendCodeResult {
-        let configuration = try self.currentCloudServiceConfiguration()
-        let result = try await self.cloudRuntime.sendCode(email: email, configuration: configuration)
-        self.globalErrorMessage = ""
-        return result
-    }
-
-    func verifyCloudOtp(challenge: CloudOtpChallenge, code: String) async throws -> CloudVerifiedAuthContext {
-        let configuration = try self.currentCloudServiceConfiguration()
-        self.globalErrorMessage = ""
-        return try await self.cloudRuntime.verifyCode(
-            challenge: challenge,
-            code: code,
-            configuration: configuration
-        )
-    }
-
-    func prepareCloudLink(verifiedContext: CloudVerifiedAuthContext) async throws -> CloudWorkspaceLinkContext {
-        let guestUpgradeMode = try await self.prepareGuestUpgradeModeIfNeeded(verifiedContext: verifiedContext)
-        let account = try await self.cloudRuntime.fetchCloudAccount(verifiedContext: verifiedContext)
-        try self.resetLocalStateIfLinkedUserDiffers(nextUserId: account.userId)
-
-        self.globalErrorMessage = ""
-        return CloudWorkspaceLinkContext(
-            userId: account.userId,
-            email: account.email,
-            apiBaseUrl: verifiedContext.apiBaseUrl,
-            credentials: verifiedContext.credentials,
-            workspaces: account.workspaces,
-            guestUpgradeMode: guestUpgradeMode
-        )
-    }
-
-    func completeCloudLink(
-        linkContext: CloudWorkspaceLinkContext,
-        selection: CloudWorkspaceLinkSelection
-    ) async throws {
-        _ = try await self.cloudRuntime.runWorkspaceCompletion { [weak self] in
-            guard let self else {
-                throw LocalStoreError.uninitialized("Flashcards store is unavailable")
-            }
-
-            guard let workspace = self.workspace else {
-                throw LocalStoreError.uninitialized("Workspace is unavailable")
-            }
-
-            let linkedWorkspace = try await self.cloudRuntime.selectOrCreateWorkspace(
-                linkContext: linkContext,
-                selection: selection,
-                localWorkspaceName: workspace.name
-            )
-
-            if try await self.shouldValidateEmptyRemoteWorkspaceBeforeBootstrap() {
-                let cloudSyncService = try requireCloudSyncService(cloudSyncService: self.dependencies.cloudSyncService)
-                let cloudSettings = try requireCloudSettings(cloudSettings: self.cloudSettings)
-                let isWorkspaceEmpty = try await cloudSyncService.isWorkspaceEmptyForBootstrap(
-                    apiBaseUrl: linkContext.apiBaseUrl,
-                    authorizationHeader: "Bearer \(linkContext.credentials.idToken)",
-                    workspaceId: linkedWorkspace.workspaceId,
-                    installationId: cloudSettings.installationId
-                )
-                if isWorkspaceEmpty == false {
-                    throw CloudBootstrapEligibilityError.remoteWorkspaceIsNotEmpty
-                }
-            }
-
-            try self.cloudRuntime.saveCredentials(credentials: linkContext.credentials)
-            let configuration = try self.currentCloudServiceConfiguration()
-            try await self.finishCloudLink(
-                linkedSession: CloudLinkedSession(
-                    userId: linkContext.userId,
-                    workspaceId: linkedWorkspace.workspaceId,
-                    email: linkContext.email,
-                    configurationMode: configuration.mode,
-                    apiBaseUrl: linkContext.apiBaseUrl,
-                    authorization: .bearer(linkContext.credentials.idToken)
-                ),
-                trigger: self.manualCloudSyncTrigger(now: Date())
-            )
-            try self.clearGuestSessionIfNeeded()
-            self.globalErrorMessage = ""
-            return linkedWorkspace
-        }
-    }
-
-    func completeGuestCloudLink(
-        linkContext: CloudWorkspaceLinkContext,
-        selection: CloudWorkspaceLinkSelection
-    ) async throws {
-        _ = try await self.cloudRuntime.runWorkspaceCompletion { [weak self] in
-            guard let self else {
-                throw LocalStoreError.uninitialized("Flashcards store is unavailable")
-            }
-
-            guard let guestSession = try self.loadGuestSessionForCurrentConfiguration() else {
-                throw LocalStoreError.uninitialized("Guest AI session is unavailable")
-            }
-
-            let guestSelection: CloudGuestUpgradeSelection
-            switch selection {
-            case .existing(let workspaceId):
-                guestSelection = .existing(workspaceId: workspaceId)
-            case .createNew:
-                guestSelection = .createNew
-            }
-
-            let linkedWorkspace = try await self.dependencies.guestCloudAuthService.completeGuestUpgrade(
-                apiBaseUrl: linkContext.apiBaseUrl,
-                bearerToken: linkContext.credentials.idToken,
-                guestToken: guestSession.guestToken,
-                selection: guestSelection
-            )
-
-            try self.cloudRuntime.saveCredentials(credentials: linkContext.credentials)
-            let configuration = try self.currentCloudServiceConfiguration()
-            try await self.finishCloudLink(
-                linkedSession: CloudLinkedSession(
-                    userId: linkContext.userId,
-                    workspaceId: linkedWorkspace.workspaceId,
-                    email: linkContext.email,
-                    configurationMode: configuration.mode,
-                    apiBaseUrl: linkContext.apiBaseUrl,
-                    authorization: .bearer(linkContext.credentials.idToken)
-                ),
-                trigger: self.manualCloudSyncTrigger(now: Date())
-            )
-            try self.clearGuestSessionIfNeeded()
-            self.globalErrorMessage = ""
-            return linkedWorkspace
-        }
-    }
-
-    func logoutCloudAccount() throws {
-        try self.resetLocalStateForCloudIdentityChange()
-    }
-
-    func applyUITestResetState(resetState: FlashcardsUITestResetState) throws {
-        try self.resetLocalStateForCloudIdentityChange()
-
-        switch resetState {
-        case .localGuest:
-            return
-        case .localGuestSeededManualReviewCard:
-            try self.seedUITestCard(
-                frontText: FlashcardsUITestSeedData.manualReviewFrontText,
-                backText: FlashcardsUITestSeedData.manualReviewBackText,
-                tags: []
-            )
-        case .localGuestSeededAIReviewCard:
-            try self.seedUITestCard(
-                frontText: FlashcardsUITestSeedData.aiReviewFrontText,
-                backText: FlashcardsUITestSeedData.aiReviewBackText,
-                tags: [FlashcardsUITestSeedData.aiReviewTag]
-            )
-        }
-    }
-
-    private func seedUITestCard(frontText: String, backText: String, tags: [String]) throws {
-        try self.saveCard(
-            input: CardEditorInput(
-                frontText: frontText,
-                backText: backText,
-                tags: tags,
-                effortLevel: .medium
-            ),
-            editingCardId: nil
-        )
-    }
-
-    func beginAccountDeletion() {
-        self.userDefaults.set(true, forKey: accountDeletionPendingUserDefaultsKey)
-        self.accountDeletionState = .inProgress
-        Task { @MainActor in
-            await self.runPendingAccountDeletion()
-        }
-    }
-
-    func retryPendingAccountDeletion() {
-        self.accountDeletionState = .inProgress
-        Task { @MainActor in
-            await self.runPendingAccountDeletion()
-        }
-    }
-
-    func resumePendingAccountDeletionIfNeeded() async {
-        guard self.userDefaults.bool(forKey: accountDeletionPendingUserDefaultsKey) else {
-            return
-        }
-
-        self.accountDeletionState = .inProgress
-        await self.runPendingAccountDeletion()
-    }
-
-    func dismissAccountDeletionSuccessMessage() {
-        self.accountDeletionSuccessMessage = nil
-    }
-
-    private func manualCloudSyncTrigger(now: Date) -> CloudSyncTrigger {
-        CloudSyncTrigger(
-            source: .manualSyncNow,
-            now: now,
-            extendsFastPolling: false,
-            allowsVisibleChangeBanner: false,
-            surfacesGlobalErrorMessage: true
-        )
-    }
-
-    func updateCurrentVisibleTab(tab: AppTab) {
-        self.currentVisibleTab = tab
-    }
-
-    func syncCloudNow(trigger: CloudSyncTrigger) async throws {
-        if case .blocked(let message) = self.syncStatus {
-            throw LocalStoreError.validation(message)
-        }
-        if self.cloudRuntime.activeCloudSession() == nil {
-            if self.cloudSettings?.cloudState == .guest {
-                let restoredGuestSession = try await self.restoreGuestCloudSessionIfNeeded(trigger: trigger)
-                if restoredGuestSession.didRunSync {
-                    return
-                }
-            } else {
-                try await self.restoreCloudLinkFromStoredCredentials(trigger: trigger)
-                return
-            }
-        }
-
-        guard let activeSession = self.cloudRuntime.activeCloudSession() else {
-            throw LocalStoreError.uninitialized("Cloud session is unavailable")
-        }
-
-        let isGuestSession = activeSession.authorization.isGuest
-        let failureStateCloudState = self.cloudSettings?.cloudState
-
-        self.syncStatus = .syncing
-        do {
-            let syncResult: CloudSyncResult
-            if isGuestSession {
-                syncResult = try await self.runLinkedSync(linkedSession: activeSession)
-            } else {
-                syncResult = try await self.withAuthenticatedCloudSession { session in
-                    try await self.runLinkedSync(linkedSession: session)
-                }
-            }
-            let now = Date()
-            try await self.applySyncResultWithoutBlockingReset(
-                syncResult: syncResult,
-                now: now,
-                trigger: trigger
-            )
-        } catch {
-            self.syncStatus = self.syncStatusForCloudFailure(
-                error: error,
-                fallbackCloudState: failureStateCloudState
-            )
-            if trigger.surfacesGlobalErrorMessage {
-                self.globalErrorMessage = Flashcards.errorMessage(error: error)
-            }
-            throw error
-        }
-    }
-
-    func syncCloudIfLinked(trigger: CloudSyncTrigger) async {
-        if self.isCloudSyncBlocked {
-            return
-        }
-        if self.userDefaults.bool(forKey: accountDeletionPendingUserDefaultsKey) {
-            await self.resumePendingAccountDeletionIfNeeded()
-            return
-        }
-
-        do {
-            let reconciliationOutcome = try await self.reconcilePersistedCloudStateBeforeSync(trigger: trigger)
-            let hasStoredCredentials: Bool
-            let hasStoredGuestSession: Bool
-            switch reconciliationOutcome {
-            case .continueSync(let resolvedHasStoredCredentials, let resolvedHasStoredGuestSession):
-                hasStoredCredentials = resolvedHasStoredCredentials
-                hasStoredGuestSession = resolvedHasStoredGuestSession
-            case .stopSync:
-                return
-            }
-
-            if try await self.cloudRuntime.waitForActiveCloudLinkTransitionIfNeeded() {
-                return
-            }
-
-            if self.cloudRuntime.activeCloudSession() == nil
-                && hasStoredCredentials == false
-                && hasStoredGuestSession == false {
-                if self.cloudSettings?.cloudState == .linked {
-                    try self.logoutCloudAccount()
-                }
-
-                return
-            }
-
-            try await self.syncCloudNow(trigger: trigger)
-        } catch {
-            if self.isCloudAccountDeletedError(error) {
-                self.handleRemoteAccountDeletedCleanup()
-                return
-            }
-
-            if trigger.surfacesGlobalErrorMessage {
-                self.globalErrorMessage = Flashcards.errorMessage(error: error)
-            }
-        }
-    }
-
-    private func reconcilePersistedCloudStateBeforeSync(
-        trigger: CloudSyncTrigger
-    ) async throws -> PersistedCloudStateReconciliationOutcome {
-        let hasStoredCredentials = try self.cloudRuntime.loadCredentials() != nil
-        let hasStoredGuestSession = try self.loadGuestSessionForCurrentConfiguration() != nil
-        guard let cloudState = self.cloudSettings?.cloudState else {
-            return .continueSync(
-                hasStoredCredentials: hasStoredCredentials,
-                hasStoredGuestSession: hasStoredGuestSession
-            )
-        }
-
-        switch cloudState {
-        case .linked:
-            if hasStoredCredentials {
-                return .continueSync(
-                    hasStoredCredentials: hasStoredCredentials,
-                    hasStoredGuestSession: hasStoredGuestSession
-                )
-            }
-
-            try self.resetLocalStateForCloudIdentityChange()
-            self.globalErrorMessage = ""
-            return .stopSync
-        case .guest:
-            let hasActiveGuestSession = self.cloudRuntime.activeCloudSession()?.authorization.isGuest == true
-            if hasStoredGuestSession || hasActiveGuestSession {
-                return .continueSync(
-                    hasStoredCredentials: hasStoredCredentials,
-                    hasStoredGuestSession: hasStoredGuestSession
-                )
-            }
-
-            try self.resetLocalStateForCloudIdentityChange()
-            self.globalErrorMessage = ""
-            return .stopSync
-        case .disconnected, .linkingReady:
-            if hasStoredGuestSession && hasStoredCredentials == false {
-                _ = try await self.restoreGuestCloudSessionIfNeeded(trigger: trigger)
-                self.globalErrorMessage = ""
-                return .stopSync
-            }
-
-            if hasStoredCredentials && hasStoredGuestSession == false {
-                let configuration = try self.currentCloudServiceConfiguration()
-                if try self.isAuthenticatedSilentRestoreEligible(
-                    configuration: configuration,
-                    hasStoredCredentials: hasStoredCredentials,
-                    hasStoredGuestSession: hasStoredGuestSession
-                ) {
-                    do {
-                        try await self.restoreAuthenticatedCloudSessionAfterReinstall(
-                            configuration: configuration,
-                            trigger: trigger
-                        )
-                        self.globalErrorMessage = ""
-                        return .stopSync
-                    } catch {
-                        if self.shouldResetLocalStateAfterAuthenticatedSilentRestoreFailure(error: error) {
-                            try self.resetLocalStateForCloudIdentityChange()
-                            self.globalErrorMessage = ""
-                            return .stopSync
-                        }
-
-                        throw error
-                    }
-                }
-
-                try self.cloudRuntime.clearCredentials()
-                self.globalErrorMessage = ""
-                return .stopSync
-            }
-
-            if hasStoredCredentials && hasStoredGuestSession {
-                try self.cloudRuntime.clearCredentials()
-                try self.dependencies.guestCredentialStore.clearGuestSession()
-                self.globalErrorMessage = ""
-                return .stopSync
-            }
-
-            return .continueSync(
-                hasStoredCredentials: hasStoredCredentials,
-                hasStoredGuestSession: hasStoredGuestSession
-            )
-        }
-    }
-
-    func cloudSessionForAI() async throws -> CloudLinkedSession {
-        if case .blocked(let message) = self.syncStatus {
-            throw LocalStoreError.validation(message)
-        }
-        if self.cloudSettings?.cloudState == .linked {
-            return try await self.prepareAuthenticatedCloudSessionForAI()
-        }
-
-        return try await self.prepareGuestCloudSessionForAI()
-    }
-
-    func warmUpAuthenticatedCloudSessionForAI() async {
-        guard self.cloudSettings?.cloudState == .linked else {
-            return
-        }
-
-        do {
-            _ = try await self.prepareAuthenticatedCloudSessionForAI()
-        } catch {
-            logFlashcardsError(
-                domain: "chat",
-                action: "ai_chat_session_warmup_failed",
-                metadata: [
-                    "message": Flashcards.errorMessage(error: error),
-                ]
-            )
-        }
-    }
-
-    private func loadOrCreateGuestCloudSession() async throws -> CloudLinkedSession {
-        let storedGuestSession: StoredGuestCloudSession
-        if let existingGuestSession = try self.loadGuestSessionForCurrentConfiguration() {
-            storedGuestSession = existingGuestSession
-        } else {
-            // After logout/account deletion the stored guest session is gone and
-            // the local installation id has already been regenerated. Creating
-            // a session here intentionally starts a brand new guest identity.
-            let configuration = try self.currentCloudServiceConfiguration()
-            let createdGuestSession = try await self.dependencies.guestCloudAuthService.createGuestSession(
-                apiBaseUrl: configuration.apiBaseUrl,
-                configurationMode: configuration.mode
-            )
-            try self.dependencies.guestCredentialStore.saveGuestSession(session: createdGuestSession)
-            storedGuestSession = createdGuestSession
-        }
-
-        return CloudLinkedSession(
-            userId: storedGuestSession.userId,
-            workspaceId: storedGuestSession.workspaceId,
-            email: nil,
-            configurationMode: storedGuestSession.configurationMode,
-            apiBaseUrl: storedGuestSession.apiBaseUrl,
-            authorization: .guest(storedGuestSession.guestToken)
-        )
-    }
-
-    private func applyGuestCloudStateBeforeReload(
-        database: LocalDatabase,
-        session: CloudLinkedSession
-    ) throws {
-        try database.updateCloudSettings(
-            cloudState: .guest,
-            linkedUserId: session.userId,
-            linkedWorkspaceId: session.workspaceId,
-            activeWorkspaceId: session.workspaceId,
-            linkedEmail: nil
-        )
-    }
-
-    private func prepareGuestCloudSessionForAI() async throws -> CloudLinkedSession {
-        let restoredGuestSession = try await self.restoreGuestCloudSessionIfNeeded(
-            trigger: CloudSyncTrigger(
-                source: .manualSyncNow,
-                now: Date(),
-                extendsFastPolling: false,
-                allowsVisibleChangeBanner: false,
-                surfacesGlobalErrorMessage: true
-            )
-        )
-        return restoredGuestSession.session
-    }
-
-    private func restoreGuestCloudSessionIfNeeded(
-        trigger: CloudSyncTrigger
-    ) async throws -> (session: CloudLinkedSession, didRunSync: Bool) {
-        let guestSession = try await self.loadOrCreateGuestCloudSession()
-        let isAlreadyGuestLinked = self.cloudSettings?.cloudState == .guest
-            && self.workspace?.workspaceId == guestSession.workspaceId
-            && self.cloudSettings?.linkedUserId == guestSession.userId
-
-        if isAlreadyGuestLinked {
-            self.cloudRuntime.setActiveCloudSession(linkedSession: guestSession)
-            return (guestSession, false)
-        }
-
-        try await self.finishCloudLink(linkedSession: guestSession, trigger: trigger)
-        return (guestSession, true)
-    }
-
-    private func prepareGuestUpgradeModeIfNeeded(
-        verifiedContext: CloudVerifiedAuthContext
-    ) async throws -> CloudGuestUpgradeMode? {
-        guard self.cloudSettings?.cloudState == .guest else {
-            return nil
-        }
-        guard let guestSession = try self.loadGuestSessionForCurrentConfiguration() else {
-            return nil
-        }
-
-        return try await self.dependencies.guestCloudAuthService.prepareGuestUpgrade(
-            apiBaseUrl: verifiedContext.apiBaseUrl,
-            bearerToken: verifiedContext.credentials.idToken,
-            guestToken: guestSession.guestToken
-        )
-    }
-
-    private func clearGuestSessionIfNeeded() throws {
-        if try self.dependencies.guestCredentialStore.loadGuestSession() != nil {
-            try self.dependencies.guestCredentialStore.clearGuestSession()
-        }
-    }
-
-    func listAgentApiKeys() async throws -> (connections: [AgentApiKeyConnection], instructions: String) {
-        return try await self.withAuthenticatedCloudSession { session in
-            let cloudSyncService = try requireCloudSyncService(cloudSyncService: self.dependencies.cloudSyncService)
-            return try await cloudSyncService.listAgentApiKeys(
-                apiBaseUrl: session.apiBaseUrl,
-                bearerToken: session.bearerToken
-            )
-        }
-    }
-
-    func revokeAgentApiKey(connectionId: String) async throws -> (connection: AgentApiKeyConnection, instructions: String) {
-        return try await self.withAuthenticatedCloudSession { session in
-            let cloudSyncService = try requireCloudSyncService(cloudSyncService: self.dependencies.cloudSyncService)
-            return try await cloudSyncService.revokeAgentApiKey(
-                apiBaseUrl: session.apiBaseUrl,
-                bearerToken: session.bearerToken,
-                connectionId: connectionId
-            )
-        }
-    }
-
-    func listLinkedWorkspaces() async throws -> [CloudWorkspaceSummary] {
-        guard self.cloudSettings?.cloudState == .linked else {
-            throw LocalStoreError.validation("Workspace switching is available only for linked cloud workspaces")
-        }
-
-        if self.cloudRuntime.activeCloudSession() == nil {
-            try await self.restoreCloudLinkFromStoredCredentials(trigger: self.manualCloudSyncTrigger(now: Date()))
-        }
-
-        return try await self.withAuthenticatedCloudSession { session in
-            let cloudSyncService = try requireCloudSyncService(cloudSyncService: self.dependencies.cloudSyncService)
-            let account = try await cloudSyncService.fetchCloudAccount(
-                apiBaseUrl: session.apiBaseUrl,
-                bearerToken: session.bearerToken
-            )
-            return account.workspaces
-        }
-    }
-
-    func switchLinkedWorkspace(selection: CloudWorkspaceLinkSelection) async throws {
-        guard self.cloudSettings?.cloudState == .linked else {
-            throw LocalStoreError.validation("Workspace switching is available only for linked cloud workspaces")
-        }
-
-        if self.cloudRuntime.activeCloudSession() == nil {
-            try await self.restoreCloudLinkFromStoredCredentials(trigger: self.manualCloudSyncTrigger(now: Date()))
-        }
-
-        let currentWorkspaceId = self.workspace?.workspaceId
-        let selectedWorkspace = try await self.withAuthenticatedCloudSession { session in
-            let cloudSyncService = try requireCloudSyncService(cloudSyncService: self.dependencies.cloudSyncService)
-            switch selection {
-            case .existing(let workspaceId):
-                return try await cloudSyncService.selectWorkspace(
-                    apiBaseUrl: session.apiBaseUrl,
-                    bearerToken: session.bearerToken,
-                    workspaceId: workspaceId
-                )
-            case .createNew:
-                return try await cloudSyncService.createWorkspace(
-                    apiBaseUrl: session.apiBaseUrl,
-                    bearerToken: session.bearerToken,
-                    name: "Personal"
-                )
-            }
-        }
-
-        if currentWorkspaceId == selectedWorkspace.workspaceId {
-            return
-        }
-
-        let activeSession = try await self.withAuthenticatedCloudSession { session in
-            CloudLinkedSession(
-                userId: session.userId,
-                workspaceId: selectedWorkspace.workspaceId,
-                email: session.email,
-                configurationMode: session.configurationMode,
-                apiBaseUrl: session.apiBaseUrl,
-                authorization: session.authorization
-            )
-        }
-
-        self.cloudRuntime.cancelForWorkspaceSwitch()
-        self.prepareWorkspaceScopedStateForSwitch(nextWorkspaceId: selectedWorkspace.workspaceId)
-        let database = try requireLocalDatabase(database: self.database)
-        try database.switchActiveWorkspace(
-            workspace: selectedWorkspace,
-            linkedSession: activeSession
-        )
-        self.cloudRuntime.setActiveCloudSession(linkedSession: activeSession)
-        try self.reload()
-        self.syncStatus = .syncing
-
-        do {
-            let syncResult = try await self.runLinkedSync(linkedSession: activeSession)
-            try await self.applySyncResultWithoutBlockingReset(
-                syncResult: syncResult,
-                now: Date(),
-                trigger: self.manualCloudSyncTrigger(now: Date())
-            )
-        } catch {
-            self.syncStatus = self.transitionSyncStatusForCloudFailure(error: error)
-            self.globalErrorMessage = Flashcards.errorMessage(error: error)
-            throw error
-        }
-    }
-
-    func renameCurrentWorkspace(name: String) async throws {
-        guard self.cloudSettings?.cloudState == .linked else {
-            throw LocalStoreError.validation("Workspace rename is available only for linked cloud workspaces")
-        }
-
-        if self.cloudRuntime.activeCloudSession() == nil {
-            try await self.restoreCloudLinkFromStoredCredentials(trigger: self.manualCloudSyncTrigger(now: Date()))
-        }
-
-        let trimmedName = name.trimmingCharacters(in: .whitespacesAndNewlines)
-        if trimmedName.isEmpty {
-            throw LocalStoreError.validation("Workspace name is required")
-        }
-
-        let workspaceId = try requireWorkspaceId(workspace: self.workspace)
-        let renamedWorkspace = try await self.withAuthenticatedCloudSession { session in
-            let cloudSyncService = try requireCloudSyncService(cloudSyncService: self.dependencies.cloudSyncService)
-            return try await cloudSyncService.renameWorkspace(
-                apiBaseUrl: session.apiBaseUrl,
-                bearerToken: session.bearerToken,
-                workspaceId: workspaceId,
-                name: trimmedName
-            )
-        }
-
-        let database = try requireLocalDatabase(database: self.database)
-        _ = try database.updateWorkspaceName(workspaceId: workspaceId, name: renamedWorkspace.name)
-        try self.reload()
-        self.globalErrorMessage = ""
-    }
-
-    func loadCurrentWorkspaceDeletePreview() async throws -> CloudWorkspaceDeletePreview {
-        guard self.cloudSettings?.cloudState == .linked else {
-            throw LocalStoreError.validation("Workspace deletion is available only for linked cloud workspaces")
-        }
-
-        if self.cloudRuntime.activeCloudSession() == nil {
-            try await self.restoreCloudLinkFromStoredCredentials(trigger: self.manualCloudSyncTrigger(now: Date()))
-        }
-
-        let workspaceId = try requireWorkspaceId(workspace: self.workspace)
-        return try await self.withAuthenticatedCloudSession { session in
-            let cloudSyncService = try requireCloudSyncService(cloudSyncService: self.dependencies.cloudSyncService)
-            return try await cloudSyncService.loadWorkspaceDeletePreview(
-                apiBaseUrl: session.apiBaseUrl,
-                bearerToken: session.bearerToken,
-                workspaceId: workspaceId
-            )
-        }
-    }
-
-    func deleteCurrentWorkspace(confirmationText: String) async throws {
-        guard self.cloudSettings?.cloudState == .linked else {
-            throw LocalStoreError.validation("Workspace deletion is available only for linked cloud workspaces")
-        }
-
-        if self.cloudRuntime.activeCloudSession() == nil {
-            try await self.restoreCloudLinkFromStoredCredentials(trigger: self.manualCloudSyncTrigger(now: Date()))
-        }
-
-        let localWorkspaceId = try requireWorkspaceId(workspace: self.workspace)
-        self.syncStatus = .syncing
-
-        do {
-            let deleteResult = try await self.withAuthenticatedCloudSession { session in
-                let cloudSyncService = try requireCloudSyncService(cloudSyncService: self.dependencies.cloudSyncService)
-                let response = try await cloudSyncService.deleteWorkspace(
-                    apiBaseUrl: session.apiBaseUrl,
-                    bearerToken: session.bearerToken,
-                    workspaceId: localWorkspaceId,
-                    confirmationText: confirmationText
-                )
-                return (session, response)
-            }
-
-            let replacementSession = CloudLinkedSession(
-                userId: deleteResult.0.userId,
-                workspaceId: deleteResult.1.workspace.workspaceId,
-                email: deleteResult.0.email,
-                configurationMode: deleteResult.0.configurationMode,
-                apiBaseUrl: deleteResult.0.apiBaseUrl,
-                authorization: deleteResult.0.authorization
-            )
-            let database = try requireLocalDatabase(database: self.database)
-            self.cloudRuntime.cancelForWorkspaceSwitch()
-            self.prepareWorkspaceScopedStateForSwitch(nextWorkspaceId: replacementSession.workspaceId)
-            try database.replaceLocalWorkspaceAfterRemoteDelete(
-                localWorkspaceId: localWorkspaceId,
-                replacementWorkspace: deleteResult.1.workspace,
-                linkedSession: replacementSession
-            )
-            self.cloudRuntime.setActiveCloudSession(linkedSession: replacementSession)
-            let syncResult = try await self.runLinkedSync(linkedSession: replacementSession)
-            try await self.applySyncResultWithoutBlockingReset(
-                syncResult: syncResult,
-                now: Date(),
-                trigger: self.manualCloudSyncTrigger(now: Date())
-            )
-        } catch {
-            self.syncStatus = self.transitionSyncStatusForCloudFailure(error: error)
-            self.globalErrorMessage = Flashcards.errorMessage(error: error)
-            throw error
-        }
-    }
-
-    func finishCloudLink(linkedSession: CloudLinkedSession, trigger: CloudSyncTrigger) async throws {
-        try await self.cloudRuntime.runCloudLinkTransition { [weak self] in
-            guard let self else {
-                throw LocalStoreError.uninitialized("Flashcards store is unavailable")
-            }
-
-            try await self.performCloudLink(linkedSession: linkedSession, trigger: trigger)
-        }
-    }
-
-    private func performCloudLink(linkedSession: CloudLinkedSession, trigger: CloudSyncTrigger) async throws {
-        if self.cloudSettings?.cloudState == .linked
-            && self.cloudSettings?.linkedUserId == linkedSession.userId {
-            let database = try requireLocalDatabase(database: self.database)
-            if self.workspace?.workspaceId == linkedSession.workspaceId {
-                try database.updateCloudSettings(
-                    cloudState: .linked,
-                    linkedUserId: linkedSession.userId,
-                    linkedWorkspaceId: linkedSession.workspaceId,
-                    activeWorkspaceId: linkedSession.workspaceId,
-                    linkedEmail: linkedSession.email
-                )
-                try self.reload()
-                try await self.performSameWorkspaceCloudRestore(linkedSession: linkedSession, trigger: trigger)
-            } else {
-                try await self.performActiveWorkspaceCloudRestore(linkedSession: linkedSession, trigger: trigger)
-            }
-            return
-        }
-
-        let context = try requireLocalMutationContext(database: self.database, workspace: self.workspace)
-
-        self.syncStatus = .syncing
-        var didCompleteLocalLink = false
-        do {
-            let remoteWorkspaceIsEmpty = try await self.isLinkedWorkspaceEmptyForBootstrap(
-                linkedSession: linkedSession
-            )
-            let migrationKind = remoteWorkspaceIsEmpty ? "preserve_local_data" : "replace_local_shell"
-            logCloudFlowPhase(
-                phase: .linkLocalWorkspace,
-                outcome: "start",
-                workspaceId: linkedSession.workspaceId,
-                installationId: self.cloudSettings?.installationId,
-                sourceWorkspaceId: context.workspaceId,
-                targetWorkspaceId: linkedSession.workspaceId,
-                migrationKind: migrationKind,
-                remoteWorkspaceIsEmpty: remoteWorkspaceIsEmpty
-            )
-            try context.database.migrateLocalWorkspaceToLinkedWorkspace(
-                localWorkspaceId: context.workspaceId,
-                linkedSession: linkedSession,
-                remoteWorkspaceIsEmpty: remoteWorkspaceIsEmpty
-            )
-            if linkedSession.authorization.isGuest {
-                try self.applyGuestCloudStateBeforeReload(
-                    database: context.database,
-                    session: linkedSession
-                )
-            }
-
-            self.cloudRuntime.setActiveCloudSession(linkedSession: linkedSession)
-            try self.reload()
-            didCompleteLocalLink = true
-            logCloudFlowPhase(
-                phase: .linkLocalWorkspace,
-                outcome: "success",
-                workspaceId: linkedSession.workspaceId,
-                installationId: self.cloudSettings?.installationId,
-                sourceWorkspaceId: context.workspaceId,
-                targetWorkspaceId: linkedSession.workspaceId,
-                migrationKind: migrationKind,
-                remoteWorkspaceIsEmpty: remoteWorkspaceIsEmpty
-            )
-            let syncResult = try await self.runLinkedSync(linkedSession: linkedSession)
-            try await self.applySyncResultWithoutBlockingReset(
-                syncResult: syncResult,
-                now: Date(),
-                trigger: trigger
-            )
-            self.userDefaults.removeObject(forKey: pendingCloudServerBootstrapUserDefaultsKey)
-            logCloudFlowPhase(
-                phase: .linkedSync,
-                outcome: "success",
-                workspaceId: linkedSession.workspaceId,
-                installationId: self.cloudSettings?.installationId
-            )
-            try self.reload()
-        } catch {
-            if didCompleteLocalLink == false {
-                logCloudFlowPhase(
-                    phase: .linkLocalWorkspace,
-                    outcome: "failure",
-                    workspaceId: linkedSession.workspaceId,
-                    installationId: self.cloudSettings?.installationId,
-                    sourceWorkspaceId: context.workspaceId,
-                    targetWorkspaceId: linkedSession.workspaceId,
-                    errorMessage: Flashcards.errorMessage(error: error)
-                )
-            }
-            logCloudFlowPhase(
-                phase: .linkedSync,
-                outcome: "failure",
-                workspaceId: linkedSession.workspaceId,
-                installationId: self.cloudSettings?.installationId,
-                errorMessage: Flashcards.errorMessage(error: error)
-            )
-            self.syncStatus = self.transitionSyncStatusForCloudFailure(error: error)
-            if trigger.surfacesGlobalErrorMessage {
-                self.globalErrorMessage = Flashcards.errorMessage(error: error)
-            }
-            throw error
-        }
-    }
-
-    /**
-     Restores a cloud session for the already-linked local workspace without
-     resetting review UI state. This keeps the locally rendered card visible
-     unless the sync result produces an actual review data change.
-     */
-    private func performSameWorkspaceCloudRestore(
-        linkedSession: CloudLinkedSession,
-        trigger: CloudSyncTrigger
-    ) async throws {
-        self.syncStatus = .syncing
-
-        do {
-            self.cloudRuntime.setActiveCloudSession(linkedSession: linkedSession)
-            let syncResult = try await self.runLinkedSync(linkedSession: linkedSession)
-            try await self.applySyncResultWithoutBlockingReset(
-                syncResult: syncResult,
-                now: Date(),
-                trigger: trigger
-            )
-            self.userDefaults.removeObject(forKey: pendingCloudServerBootstrapUserDefaultsKey)
-        } catch {
-            logCloudFlowPhase(
-                phase: .linkedSync,
-                outcome: "failure",
-                workspaceId: linkedSession.workspaceId,
-                installationId: self.cloudSettings?.installationId,
-                errorMessage: Flashcards.errorMessage(error: error)
-            )
-            self.syncStatus = self.transitionSyncStatusForCloudFailure(error: error)
-            if trigger.surfacesGlobalErrorMessage {
-                self.globalErrorMessage = Flashcards.errorMessage(error: error)
-            }
-            throw error
-        }
-    }
-
-    private func performActiveWorkspaceCloudRestore(
-        linkedSession: CloudLinkedSession,
-        trigger: CloudSyncTrigger
-    ) async throws {
-        let database = try requireLocalDatabase(database: self.database)
-        let cachedWorkspace = try database.loadCachedWorkspaces().first { workspace in
-            workspace.workspaceId == linkedSession.workspaceId
-        }
-        let workspaceSummary = CloudWorkspaceSummary(
-            workspaceId: linkedSession.workspaceId,
-            name: cachedWorkspace?.name ?? "Personal",
-            createdAt: cachedWorkspace?.createdAt ?? nowIsoTimestamp(),
-            isSelected: true
-        )
-
-        self.cloudRuntime.cancelForWorkspaceSwitch()
-        self.prepareWorkspaceScopedStateForSwitch(nextWorkspaceId: linkedSession.workspaceId)
-        try database.switchActiveWorkspace(workspace: workspaceSummary, linkedSession: linkedSession)
-        self.cloudRuntime.setActiveCloudSession(linkedSession: linkedSession)
-        try self.reload()
-        try await self.performSameWorkspaceCloudRestore(linkedSession: linkedSession, trigger: trigger)
-    }
-
-    /**
-     Applies sync side effects through diff-aware bootstrap and review
-     reconciliation so no-op syncs do not trigger a blocking review reload.
-     */
-    private func applySyncResultWithoutBlockingReset(
-        syncResult: CloudSyncResult,
-        now: Date,
-        trigger: CloudSyncTrigger
-    ) async throws {
-        let bootstrapRefreshOutcome = try self.refreshBootstrapSnapshotWithoutReset(now: now)
-        let didRefreshReviewState: Bool
-        if syncResult.reviewDataChanged {
-            let reviewRefreshMode: ReviewRefreshMode
-            if trigger.allowsVisibleChangeBanner {
-                reviewRefreshMode = .backgroundReconcileWithVisibleChangeBanner
-            } else {
-                reviewRefreshMode = .backgroundReconcileSilently
-            }
-            didRefreshReviewState = try await self.refreshReviewState(
-                now: now,
-                mode: reviewRefreshMode
-            )
-        } else {
-            didRefreshReviewState = false
-        }
-        if trigger.allowsVisibleChangeBanner {
-            self.enqueueBackgroundSyncVisibleChangeBannerIfNeeded(
-                bootstrapRefreshOutcome: bootstrapRefreshOutcome
-            )
-        }
-        if bootstrapRefreshOutcome.didChange || didRefreshReviewState {
-            self.localReadVersion += 1
-        }
-        self.lastSuccessfulCloudSyncAt = nowIsoTimestamp()
-        self.syncStatus = .idle
-        self.globalErrorMessage = ""
-    }
-
-    func refreshCloudCredentials(forceRefresh: Bool) async throws -> StoredCloudCredentials {
-        let configuration = try self.currentCloudServiceConfiguration()
-        return try await self.cloudRuntime.refreshCloudCredentials(
-            forceRefresh: forceRefresh,
-            configuration: configuration,
-            now: Date()
-        )
-    }
-
-    func prepareAuthenticatedCloudSessionForAI() async throws -> CloudLinkedSession {
-        return try await self.cloudRuntime.prepareAuthenticatedCloudSessionForAI(
-            restoreCloudLink: { [weak self] in
-                guard let self else {
-                    throw LocalStoreError.uninitialized("Flashcards store is unavailable")
-                }
-
-                try await self.restoreCloudLinkFromStoredCredentials(trigger: self.manualCloudSyncTrigger(now: Date()))
-            },
-            resolveSession: { [weak self] in
-                guard let self else {
-                    throw LocalStoreError.uninitialized("Flashcards store is unavailable")
-                }
-
-                return try await self.withAuthenticatedCloudSession { session in
-                    session
-                }
-            }
-        )
-    }
-
-    func restoreCloudLinkFromStoredCredentials(trigger: CloudSyncTrigger) async throws {
-        try await self.cloudRuntime.runCloudLinkTransition { [weak self] in
-            guard let self else {
-                throw LocalStoreError.uninitialized("Flashcards store is unavailable")
-            }
-
-            try await self.performRestoreCloudLinkFromStoredCredentials(trigger: trigger)
-        }
-    }
-
-    /**
-     Restores the linked cloud session from persisted credentials. When the
-     stored workspace already matches the local workspace, it reuses the
-     non-blocking restore path; otherwise it falls back to the full relink flow.
-     */
-    private func performRestoreCloudLinkFromStoredCredentials(trigger: CloudSyncTrigger) async throws {
-        let configuration = try self.currentCloudServiceConfiguration()
-
-        do {
-            let credentials = try await self.refreshCloudCredentials(forceRefresh: false)
-            if try await self.resetLocalStateIfStoredCredentialsBelongToDifferentUser(
-                credentials: credentials,
-                configuration: configuration
-            ) {
-                return
-            }
-            let linkedSession = try self.cloudRuntime.storedLinkedSession(
-                cloudSettings: self.cloudSettings,
-                configuration: configuration,
-                bearerToken: credentials.idToken
-            )
-            if self.workspace?.workspaceId == linkedSession.workspaceId {
-                try await self.performSameWorkspaceCloudRestore(linkedSession: linkedSession, trigger: trigger)
-            } else {
-                try await self.performActiveWorkspaceCloudRestore(linkedSession: linkedSession, trigger: trigger)
-            }
-            return
-        } catch {
-            if self.isCloudAccountDeletedError(error) {
-                self.handleRemoteAccountDeletedCleanup()
-                return
-            }
-
-            if self.isCloudAuthorizationError(error) == false {
-                throw error
-            }
-        }
-
-        do {
-            let refreshedCredentials = try await self.refreshCloudCredentials(forceRefresh: true)
-            if try await self.resetLocalStateIfStoredCredentialsBelongToDifferentUser(
-                credentials: refreshedCredentials,
-                configuration: configuration
-            ) {
-                return
-            }
-            let linkedSession = try self.cloudRuntime.storedLinkedSession(
-                cloudSettings: self.cloudSettings,
-                configuration: configuration,
-                bearerToken: refreshedCredentials.idToken
-            )
-            if self.workspace?.workspaceId == linkedSession.workspaceId {
-                try await self.performSameWorkspaceCloudRestore(linkedSession: linkedSession, trigger: trigger)
-            } else {
-                try await self.performActiveWorkspaceCloudRestore(linkedSession: linkedSession, trigger: trigger)
-            }
-        } catch {
-            if self.isCloudAccountDeletedError(error) {
-                self.handleRemoteAccountDeletedCleanup()
-                return
-            }
-
-            if self.isCloudAuthorizationError(error) {
-                try self.logoutCloudAccount()
-            }
-
-            throw error
-        }
-    }
-
-    func withAuthenticatedCloudSession<Result>(
-        operation: (CloudLinkedSession) async throws -> Result
-    ) async throws -> Result {
-        do {
-            let credentials = try await self.refreshCloudCredentials(forceRefresh: false)
-            let linkedSession = try self.cloudRuntime.sessionWithUpdatedBearerToken(credentials: credentials)
-            return try await operation(linkedSession)
-        } catch {
-            if self.isCloudAccountDeletedError(error) {
-                self.handleRemoteAccountDeletedCleanup()
-                throw error
-            }
-
-            if self.isCloudAuthorizationError(error) == false {
-                throw error
-            }
-        }
-
-        do {
-            let refreshedCredentials = try await self.refreshCloudCredentials(forceRefresh: true)
-            let linkedSession = try self.cloudRuntime.sessionWithUpdatedBearerToken(credentials: refreshedCredentials)
-            return try await operation(linkedSession)
-        } catch {
-            if self.isCloudAccountDeletedError(error) {
-                self.handleRemoteAccountDeletedCleanup()
-                throw error
-            }
-
-            if self.isCloudAuthorizationError(error) {
-                try self.logoutCloudAccount()
-            }
-
-            throw error
-        }
-    }
-
-    func isCloudAuthorizationError(_ error: Error) -> Bool {
-        self.cloudRuntime.isCloudAuthorizationError(error)
-    }
-
-    var isCloudSyncBlocked: Bool {
-        if case .blocked = self.syncStatus {
-            return true
-        }
-        return false
-    }
-
-    func isCloudAccountDeletedError(_ error: Error) -> Bool {
-        self.cloudRuntime.isCloudAccountDeletedError(error)
-    }
-
-    private func syncStatusForCloudFailure(
-        error: Error,
-        fallbackCloudState: CloudAccountState?
-    ) -> SyncStatus {
-        if let blockedMessage = self.blockedCloudIdentityConflictMessage(error: error) {
-            return .blocked(message: blockedMessage)
-        }
-
-        if fallbackCloudState == .linked || fallbackCloudState == .guest {
-            return .failed(message: Flashcards.errorMessage(error: error))
-        }
-
-        return .idle
-    }
-
-    private func transitionSyncStatusForCloudFailure(error: Error) -> SyncStatus {
-        if let blockedMessage = self.blockedCloudIdentityConflictMessage(error: error) {
-            return .blocked(message: blockedMessage)
-        }
-
-        return .failed(message: Flashcards.errorMessage(error: error))
-    }
-
-    private func blockedCloudIdentityConflictMessage(error: Error) -> String? {
-        guard let syncError = error as? CloudSyncError else {
-            return nil
-        }
-        guard case .invalidResponse(let details, _) = syncError else {
-            return nil
-        }
-        guard blockedCloudIdentityConflictCodes.contains(details.code ?? "") else {
-            return nil
-        }
-        return Flashcards.errorMessage(error: error)
-    }
-
-    func runPendingAccountDeletion() async {
-        guard self.isAccountDeletionRunning == false else {
-            return
-        }
-
-        self.isAccountDeletionRunning = true
-        defer {
-            self.isAccountDeletionRunning = false
-        }
-
-        do {
-            try await self.performCloudAccountDeletion()
-            try self.completeLocalAccountDeletion()
-            self.accountDeletionState = .hidden
-            self.accountDeletionSuccessMessage = "Your account has been deleted."
-        } catch {
-            if self.isCloudAccountDeletedError(error) {
-                return
-            }
-
-            self.accountDeletionState = .failed(message: Flashcards.errorMessage(error: error))
-        }
-    }
-
-    func performCloudAccountDeletion() async throws {
-        try await self.withAuthenticatedCloudSession { session in
-            let cloudSyncService = try requireCloudSyncService(cloudSyncService: self.dependencies.cloudSyncService)
-            try await cloudSyncService.deleteAccount(
-                apiBaseUrl: session.apiBaseUrl,
-                bearerToken: session.bearerToken,
-                confirmationText: accountDeletionConfirmationText
-            )
-        }
-    }
-
-    func completeLocalAccountDeletion() throws {
-        self.userDefaults.removeObject(forKey: accountDeletionPendingUserDefaultsKey)
-        try self.resetLocalStateForCloudIdentityChange()
-    }
-
-    func handleRemoteAccountDeletedCleanup() {
-        do {
-            self.userDefaults.set(true, forKey: accountDeletionPendingUserDefaultsKey)
-            try self.completeLocalAccountDeletion()
-            self.accountDeletionState = .hidden
-            self.accountDeletionSuccessMessage = "Your account has been deleted."
-        } catch {
-            self.accountDeletionState = .failed(message: Flashcards.errorMessage(error: error))
-        }
-    }
-
-    func runLinkedSync(linkedSession: CloudLinkedSession) async throws -> CloudSyncResult {
-        try await self.cloudRuntime.runLinkedSync(linkedSession: linkedSession)
-    }
-
-    func triggerCloudSyncIfLinked(trigger: CloudSyncTrigger) {
-        if trigger.extendsFastPolling {
-            self.extendCloudSyncFastPolling(now: trigger.now)
-        }
-        if self.shouldSkipImmediateCloudSyncStart(trigger: trigger) {
-            return
-        }
-        Task { @MainActor in
-            await self.syncCloudIfLinked(trigger: trigger)
-        }
-    }
-
-    private func shouldSkipImmediateCloudSyncStart(trigger: CloudSyncTrigger) -> Bool {
-        guard trigger.source.usesImmediateStartDebounce else {
-            return false
-        }
-        if let lastImmediateCloudSyncTriggerAt,
-           trigger.now.timeIntervalSince(lastImmediateCloudSyncTriggerAt) < cloudImmediateSyncDebounceIntervalSeconds {
-            return true
-        }
-
-        self.lastImmediateCloudSyncTriggerAt = trigger.now
-        return false
-    }
-
-    private func enqueueBackgroundSyncVisibleChangeBannerIfNeeded(
-        bootstrapRefreshOutcome: BootstrapSnapshotRefreshOutcome
-    ) {
-        guard self.currentVisibleTab == .cards else {
-            return
-        }
-        guard bootstrapRefreshOutcome.workspaceChanged
-            || bootstrapRefreshOutcome.cardsChanged else {
-            return
-        }
-
-        self.enqueueTransientBanner(banner: makeCardsUpdatedFromCloudBanner())
-    }
-
-    private func switchCloudServer(override: CloudServerOverride?) throws {
-        let context = try requireLocalMutationContext(database: self.database, workspace: self.workspace)
-
-        self.cloudRuntime.cancelForAccountDeletion()
-        try self.cloudRuntime.clearCredentials()
-        try self.dependencies.guestCredentialStore.clearGuestSession()
-        try context.database.clearCloudSyncState(workspaceId: context.workspaceId)
-        try context.database.updateCloudSettings(
-            cloudState: .disconnected,
-            linkedUserId: nil,
-            linkedWorkspaceId: nil,
-            activeWorkspaceId: context.workspaceId,
-            linkedEmail: nil
-        )
-
-        if let override {
-            try saveCloudServerOverride(
-                override: override,
-                userDefaults: self.userDefaults,
-                encoder: self.encoder
-            )
-        } else {
-            clearCloudServerOverride(userDefaults: self.userDefaults)
-        }
-
-        if override == nil {
-            self.userDefaults.removeObject(forKey: pendingCloudServerBootstrapUserDefaultsKey)
-        } else {
-            self.userDefaults.set(true, forKey: pendingCloudServerBootstrapUserDefaultsKey)
-        }
-        self.syncStatus = .idle
-        self.lastSuccessfulCloudSyncAt = nil
-        self.globalErrorMessage = ""
-        try self.reload()
-    }
-
-    private func shouldValidateEmptyRemoteWorkspaceBeforeBootstrap() async throws -> Bool {
-        let configuration = try self.currentCloudServiceConfiguration()
-        return configuration.mode == .custom
-            && self.userDefaults.bool(forKey: pendingCloudServerBootstrapUserDefaultsKey)
-    }
-
-    private func isLinkedWorkspaceEmptyForBootstrap(linkedSession: CloudLinkedSession) async throws -> Bool {
-        let cloudSyncService = try requireCloudSyncService(cloudSyncService: self.dependencies.cloudSyncService)
-        let cloudSettings = try requireCloudSettings(cloudSettings: self.cloudSettings)
-        return try await cloudSyncService.isWorkspaceEmptyForBootstrap(
-            apiBaseUrl: linkedSession.apiBaseUrl,
-            authorizationHeader: linkedSession.authorizationHeaderValue,
-            workspaceId: linkedSession.workspaceId,
-            installationId: cloudSettings.installationId
-        )
-    }
-}
